+++ backend/server.js
@@ -11,7 +11,7 @@ import productRoutes from './routes/productRoutes.js';
 import orderRoutes from './routes/orderRoutes.js';
 import newsletterRoutes from './routes/newsletterRoutes.js';
 import eCardRoutes from './routes/eCardRoutes.js';
-import blogRoutes from './routes/blogRoutes.js';
+import articleRoutes from './routes/articleRoutes.js';
 import educationTipRoutes from './routes/educationTipRoutes.js';
 import boardMemberRoutes from './routes/boardMemberRoutes.js';
 import welcomeWienerDog from './routes/welcomeWienerDogRoutes.js';
@@ -84,7 +84,7 @@ app.use('/api/donation', donationRoutes);
 app.use('/api/products', productRoutes);
 app.use('/api/newsletter', newsletterRoutes);
 app.use('/api/ecard', eCardRoutes);
-app.use('/api/blog', blogRoutes);
+app.use('/api/blog', articleRoutes);
 app.use('/api/board-member', boardMemberRoutes);
 app.use('/api/welcome-wiener-dog', welcomeWienerDog);
 app.use('/api/welcome-wiener-product', welcomeWienerProduct);
